+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.2
 
-* TODO: Describe initial release.
+This is second version of the package
+
+We have include 2 bar chart sample example in this update for animation bars with example
+
+We keep update our new charts example soon.